requirements.txt
httpx==0.27.2
pydantic==2.9.2
python-dotenv==1.0.1
rich==13.8.1
tenacity==9.0.0
python-telegram-bot==21.5
